Updated authoptions export for prod build
