+++ src/utils/serverUtils.ts
@@ -1,8 +1,8 @@
 'use server';
 
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
 import { AuthSession } from '@/types/auth';
 import { getServerSession } from 'next-auth/next';
+import { authOptions } from './authOptions';
 
 export const customGet = async (url: string, session: AuthSession | null) => {
   if (!session) {
